requirements.txt
streamlit==0.89.0
numpy==1.19.3
pandas==1.3.3
matplotlib==3.4.2
mysql-connector-python==8.0.26
scikit-learn==1.0
seaborn==0.11.2
pickle-mixin==1.0.2
tensorflow-cpu==2.6.0
